Account Creation
Added SQLite database, Account Creation page, log in page, and a profile
page that shows the user's information as well as all other users on the
database.

To run, make sure that apache is configured so that the document path is
set to this folder and that the cgi-bin path is correctly set.